requirements.txt
requests>=2.28.0
google-cloud-storage>=2.0.0
google-cloud-pubsub>=2.18.0
google-cloud-bigquery>=3.11.0
